requirements.txt
Django>=4.2
djangorestframework
psycopg[binary]
django-filter
drf-yasg  # API dokumentatsiya uchun (Swagger)
channels 
channels-redis
django_filters
django-modeltranslation

